Buyer/Buyer_server/requirements.txt
asyncpg==0.22.0
beautifulsoup4==4.9.3
gino==1.0.1
google==3.0.0
googleapis-common-protos==1.52.0
grpcio==1.35.0
grpcio-tools==1.35.0
protobuf==3.15.1
six==1.15.0
soupsieve==2.2
SQLAlchemy==1.3.23